General commit: ========================
Added the application-db.properties and application.properties file for the configurations.
Added the swagger configurations.
Created a database on AWS for the API to work with.
Created the User.java, Photo.java, and Shared.java classes (persistence objects).
Created the response class for delivering the result of a request.
Also ignored the application-db.properties file for highly confidential information.
Had to disable auto datasource configuration due to an error regarding the datasource url???(added it as a TODO)