Advanced commit: ========================
Users can now be added via swagger API.
At the moment passwords are being encrypted and stored as such in the database.
Added the 3 DTOs along with their respective functionality.
Now have to add the view, update, and delete parts for the users and once that's finished advance to the images part of the project.